Add Author Tags to MODELS and CONSTANTS
