+++ lib/constants/country_codes.dart
@@ -1,3 +1,9 @@
+//  Country Codes - Constant List of Supported Languages (NOTE: ADDING A LANGUAGE REQUIRES CHANGES IN MAIN.DART TOO!!!)
+//
+//  Author: NightmindOfficial
+//  Co-Author: damattl
+//  (Refactored)
+
 import 'package:linum/utilities/frontend/country_flag_generator.dart';
 
 final Map<String, String> countryFlagsToCountryCode = {
