+++ lib/constants/repeat_duration_type_enum.dart
@@ -1,0 +1,10 @@
+//  Repeat Duration Type - Enum that specifies the unit of a repeat duration
+//
+//  Author: SoTBurst
+//  Co-Author: n/a
+//
+
+enum RepeatDurationType {
+  seconds,
+  months,
+}
