+++ lib/constants/repeatable_change_type_enum.dart
@@ -1,0 +1,12 @@
+//  Repeatable Change Type - Enum that specifies all Options a user has for deleting entries that are part of a repeatable
+//
+//  Author: SoTBurst
+//  Co-Author: n/a
+//
+
+enum RepeatableChangeType {
+  all,
+  thisAndAllBefore,
+  thisAndAllAfter,
+  onlyThisOne,
+}
