+++ lib/constants/ring_colors.dart
@@ -1,3 +1,9 @@
+//  Ring Colors - Contains Colors exclusively used in the PIN Lock
+//
+//  Author: NightmindOfficial
+//  Co-Author: damattl
+//  (Refactored)
+
 import 'package:flutter/material.dart';
 
 class RingColors {
