+++ lib/models/entry_category.dart
@@ -1,3 +1,9 @@
+//  EntryCategoriy - Basic Model of a Category that can be displayed within the settings_screen.dart
+//
+//  Author: NightmindOfficial
+//  Co-Author: n/a
+//
+
 import 'package:flutter/material.dart';
 
 class EntryCategory {
