+++ lib/models/home_screen_card_data.dart
@@ -1,3 +1,9 @@
+//  HomeScreenCardData - Basic Model for what can be displayed within the home_screen_card.dart
+//
+//  Author: damattl
+//  Co-Author: NightmindOfficial
+//  (Refactored)
+
 class HomeScreenCardData {
   final num balance;
   final num income;
