--- lib/models/repeat_duration_type_enum.dart
@@ -1,4 +1,0 @@
-enum RepeatDurationType {
-  seconds,
-  months,
-}
