--- lib/models/repeatable_change_type.dart
@@ -1,6 +1,0 @@
-enum RepeatableChangeType {
-  all,
-  thisAndAllBefore,
-  thisAndAllAfter,
-  onlyThisOne,
-}
