+++ lib/models/single_balance_data.dart
@@ -1,3 +1,9 @@
+//  Single Balance Data - Model for defining information used by a card that is NOT a copy of an original repeated transaction ("Single Transacton").
+//
+//  Author: SoTBurst
+//  Co-Author: n/a //TODO @SoTBurst might be useful to train more people on this
+//
+
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:uuid/uuid.dart';
