+++ lib/providers/balance_data_provider.dart
@@ -3,8 +3,8 @@ import 'dart:developer' as dev;
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:linum/models/repeat_balance_data.dart';
-import 'package:linum/models/repeat_duration_type_enum.dart';
-import 'package:linum/models/repeatable_change_type.dart';
+import 'package:linum/constants/repeat_duration_type_enum.dart';
+import 'package:linum/constants/repeatable_change_type_enum.dart';
 import 'package:linum/models/single_balance_data.dart';
 import 'package:linum/providers/algorithm_provider.dart';
 import 'package:linum/providers/authentication_service.dart';
