+++ lib/screens/enter_screen.dart
@@ -4,7 +4,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:linum/models/dialog_action.dart';
 import 'package:linum/models/repeat_balance_data.dart';
-import 'package:linum/models/repeatable_change_type.dart';
+import 'package:linum/constants/repeatable_change_type_enum.dart';
 import 'package:linum/models/single_balance_data.dart';
 import 'package:linum/providers/balance_data_provider.dart';
 import 'package:linum/providers/enter_screen_provider.dart';
