+++ test/balance_data/repeated_balance_data_manager_test.dart
@@ -4,8 +4,8 @@ import 'dart:math' as math;
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:linum/models/repeat_balance_data.dart';
-import 'package:linum/models/repeat_duration_type_enum.dart';
-import 'package:linum/models/repeatable_change_type.dart';
+import 'package:linum/constants/repeat_duration_type_enum.dart';
+import 'package:linum/constants/repeatable_change_type_enum.dart';
 import 'package:linum/utilities/backend/date_time_calculation_functions.dart';
 import 'package:linum/utilities/backend/repeated_balance_help_functions.dart';
 import 'package:linum/utilities/balance_data/repeated_balance_data_manager.dart';
